refactored project structure (ui/components are now more organized
