+++ src/app/dashboard/page.js
@@ -1,19 +1,16 @@
 "use client";
-// migrating new dashboard from "the graph" branch
-import Button from '../../components/Button';
-import TopNav from '../../components/TopNav';
-import TimeDataView from '../../components/TimeDataView';
-import GraphDemo from '../../components/GraphDemo';
-import ShiftsGraph from '../../components/ShiftsGraph';
-import Heatmap from '../../components/Heatmap';
+import TopNav from '../../ui/TopNav';
+import TimeDataView from './components/TimeDataView';
+import ShiftsGraph from './components/ShiftsGraph';
+import Heatmap from './components/Heatmap';
 import { useEffect, useState } from "react";
 import { supabase } from '../../lib/supabaseClient';
-import ProfileCompletionPrompt from '../../components/ProfileCompletionPrompt';
-import ProfilePanel from '../../components/ProfilePanel';
-import Spreadsheet from '../../components/Spreadsheet';
+import ProfileCompletionPrompt from './components/ProfileCompletionPrompt';
+import ProfilePanel from './components/ProfilePanel';
+import Spreadsheet from './components/Spreadsheet';
 import { getUserWages } from '../../utils/wageUtils';
 import { getUserProfile } from '../../utils/profileUtils';
-import Test from '../../components/test';
+import RecentShifts from './components/RecentShifts';
 import { ShiftsProvider } from '../../context/ShiftsContext';
 
 
@@ -67,7 +64,6 @@ export default function Dashboard() {
                     setGreetingName(user.email?.split('@')[0] || 'Welcome');
                 }
             } catch (e) {
-                // non-fatal; just skip prompt on error
                 console.warn('Failed checking wages:', e?.message);
             }
         };
@@ -121,11 +117,9 @@ export default function Dashboard() {
                 userId={currentUser?.id}
             />
         )}
-            {/* Global Nav */}
-            <TopNav />
+           <TopNav />
 
-            {/* Main Content */}
-            <ShiftsProvider>
+           <ShiftsProvider>
             <main className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
                 {/* Stats Grid */}
                 <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6 mb-8">
@@ -136,21 +130,18 @@ export default function Dashboard() {
                   <h1 className="text-2xl font-bold text-text-primary">{`Welcome${greetingName ? `, ${greetingName}` : ''}`}</h1>
                   <p className="text-text-secondary">Personal Work Schedule Management</p>
                 </div>
-
-                {/* Content Grid */}
-                <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 mb-8">
+               <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 mb-8">
                     {/* Recent Shifts */}
                     <div className="bg-white p-6 rounded-lg shadow-sm border border-accent">
                         <h2 className="text-xl font-semibold text-gray-900 mb-4">Recent Shifts</h2>
                         <div className="space-y-4">
-                        <Test />
+                        <RecentShifts />
                         </div>
                     </div>
                 </div>
 
 
 
-                {/* Time Data Analytics Section with Stale-While-Revalidate Graph */}
                 <TimeDataView 
                     title="Shifts Analytics"
                     onTimeRangeChange={handleTimeRangeChange}
@@ -164,7 +155,6 @@ export default function Dashboard() {
                         className="m-4"
                     />
                 </TimeDataView>
-                {/*standard box for heatmap*/}
                 <div className="bg-white p-6 rounded-lg shadow-sm border border-accent mt-8">
                    <Heatmap />
