+++ src/app/paycheck-audit/components/MissionStatement.js
@@ -1,0 +1,11 @@
+export default function MissionStatement() {
+  return (
+    <section className="rounded-lg border border-border-light bg-surface p-6">
+      <h3 className="text-base font-semibold text-text-primary mb-2">Why this matters</h3>
+      <p className="text-sm text-text-secondary">
+        Placeholder mission statement
+      </p>
+    </section>
+  );
+}
+
