--- src/components/Pdfparser.js
@@ -1,10 +1,0 @@
-
-export default function Pdfparser() {
-
-    return (
-        <div>
-            <p>Pdfparser placeholder</p>
-        </div>
-
-    )
-}
