+++ src/ui/Button.js
@@ -50,3 +50,4 @@ export default function Button({
         </button>
     );
 }
+
